Rename to VAudioControl
