+++ frontend/src/components/VAudioTrack/VAudioTrack.vue
@@ -37,7 +37,7 @@
       </template>
 
       <template #play-pause="playPauseProps">
-        <VOldPlayPause
+        <VPlayPause
           ref="playPauseRef"
           :status="status"
           v-bind="playPauseProps"
@@ -88,7 +88,7 @@ import { defineEvent } from "~/types/emits"
 
 import type { AudioTrackClickEvent } from "~/types/events"
 
-import VOldPlayPause from "~/components/VAudioTrack/VOldPlayPause.vue"
+import VPlayPause from "~/components/VAudioTrack/VPlayPause.vue"
 import VWaveform from "~/components/VAudioTrack/VWaveform.vue"
 import VFullLayout from "~/components/VAudioTrack/layouts/VFullLayout.vue"
 import VRowLayout from "~/components/VAudioTrack/layouts/VRowLayout.vue"
@@ -104,7 +104,7 @@ import VWarningSuppressor from "~/components/VWarningSuppressor.vue"
 export default defineComponent({
   name: "VAudioTrack",
   components: {
-    VOldPlayPause,
+    VPlayPause,
     VWaveform,
     VLink,
     VWarningSuppressor,
