+++ frontend/src/components/VAudioTrack/VPlayPause.vue
@@ -1,12 +1,11 @@
 <template>
-  <VIconButton
+  <VOldIconButton
+    v-bind="$attrs"
     :tabindex="isTabbable ? 0 : -1"
-    class="play-pause"
-    :size="buttonSize"
-    variant="filled-dark"
-    :icon-props="icon === undefined ? undefined : { name: icon, size: iSize }"
+    class="play-pause flex-shrink-0 border-dark-charcoal bg-dark-charcoal text-white focus-visible:border-pink focus-visible:shadow-ring focus-visible:outline-none active:shadow-ring disabled:opacity-70"
+    :icon-props="icon === undefined ? undefined : { name: icon }"
     :label="$t(label)"
-    :connections="connections"
+    :button-props="buttonProps"
     @click.stop.prevent="handleClick"
     @mousedown="handleMouseDown"
   >
@@ -27,23 +26,17 @@
         />
       </svg>
     </template>
-  </VIconButton>
+  </VOldIconButton>
 </template>
 
 <script lang="ts">
 import { computed, defineComponent, PropType } from "vue"
 
-import {
-  AudioLayout,
-  audioLayouts,
-  AudioStatus,
-  audioStatuses,
-  statusVerbMap,
-} from "~/constants/audio"
+import { AudioLayout, AudioStatus, statusVerbMap } from "~/constants/audio"
 import { defineEvent } from "~/types/emits"
-import type { ButtonConnections } from "~/types/button"
+import type { ButtonConnections, ButtonVariant } from "~/types/button"
 
-import VIconButton from "~/components/VIconButton/VIconButton.vue"
+import VOldIconButton from "~/components/VIconButton/VOldIconButton.vue"
 
 const statusIconMap = {
   playing: "pause",
@@ -60,22 +53,13 @@ const layoutConnectionsMap: Record<AudioLayout, ButtonConnections> = {
 } as const
 
 /**
- * The mapping of play-pause control sizes to the VIconButton sizes
- * and the sizes of the contained icon.
- */
-const sizes = {
-  small: { button: "small", icon: 6 },
-  medium: { button: "large", icon: 8 },
-  large: { button: "larger", icon: 10 },
-} as const
-
-/**
  * Displays the control for switching between the playing and paused states of
  * a media file.
  */
 export default defineComponent({
   name: "VPlayPause",
-  components: { VIconButton },
+  components: { VOldIconButton },
+  inheritAttrs: false,
   model: {
     prop: "status",
     event: "toggle",
@@ -87,30 +71,16 @@ export default defineComponent({
     status: {
       type: String as PropType<AudioStatus>,
       required: true,
-      validator: (val: string) =>
-        (audioStatuses as readonly string[]).includes(val),
     },
     /**
-     * The size of the button. The size affects both the size of the button
-     * itself and the icon inside it.
-     */
-    size: {
-      type: String as PropType<"small" | "medium" | "large">,
-      required: true,
-      validator: (val: string) => ["small", "medium", "large"].includes(val),
-    },
-    /**
-     * The parent audio layout currently in use. The connections are determined
-     * by the layout and the size of the button.
+     * The parent audio layout currently in use
      */
     layout: {
       type: String as PropType<AudioLayout>,
       default: "full",
-      validator: (val: string) =>
-        (audioLayouts as readonly string[]).includes(val),
     },
     /**
-     * Whether the play-pause button can be focused by using the `Tab` key
+     * whether the play-pause button can be focused by using the `Tab` key
      */
     isTabbable: {
       type: Boolean,
@@ -133,18 +103,14 @@ export default defineComponent({
     const icon = computed(() => statusIconMap[props.status])
 
     /**
-     * Set the connections (none-rounded corners) for the button based on the layout.
+     * Sets the button variant to `plain--avoid` to manually handle focus states.
+     * Sets the connections (none-rounded corners) for the button based on the layout.
      */
-    const connections = computed(() => {
-      return props.layout === "row" && props.size === "small"
-        ? "none"
-        : layoutConnectionsMap[props.layout]
+    const buttonProps = computed(() => {
+      const variant = "plain--avoid" as ButtonVariant
+
+      return { variant, connections: layoutConnectionsMap[props.layout] }
     })
-
-    /** Convert the `play-pause` sizes to `VIconButton` sizes */
-    const buttonSize = computed(() => sizes[props.size].button)
-
-    const iSize = computed(() => sizes[props.size].icon)
 
     const handleMouseDown = (event: MouseEvent) => {
       if (!props.isTabbable) event.preventDefault() // to prevent focus
@@ -152,13 +118,10 @@ export default defineComponent({
     const handleClick = () => {
       emit("toggle", isPlaying.value || isLoading.value ? "paused" : "playing")
     }
-
     return {
       label,
       icon,
-      connections,
-      buttonSize,
-      iSize,
+      buttonProps,
       isLoading,
 
       handleClick,
